Replacing ternary operator with select in ispc which is branchless

+++ TradinatorCore/ispc/indicator_helper_ispc.h
@@ -33,10 +33,10 @@ namespace ispc { /* namespace */
 #if defined(__cplusplus) && (! defined(__ISPC_NO_EXTERN_C) || !__ISPC_NO_EXTERN_C )
 extern "C" {
 #endif // __cplusplus
-    extern void calculate_bollinger_band(double * input, double * top_output, double * sma_output, double * bottom_output, uint32_t data_size, uint32_t window_size, double standard_deviation_multiplier);
-    extern void calculate_ema(double * input, double * output, uint32_t data_size, uint32_t window_size);
-    extern void calculate_sma(double * input, double * output, uint32_t data_size, uint32_t window_size);
-    extern void calculate_wma(double * input, double * output, uint32_t data_size, uint32_t window_size);
+    extern void calculate_bollinger_band(double * input, double * top_output, double * sma_output, double * bottom_output, uint64_t data_size, uint64_t window_size, double standard_deviation_multiplier);
+    extern void calculate_ema(double * input, double * output, uint64_t data_size, uint64_t window_size);
+    extern void calculate_sma(double * input, double * output, uint64_t data_size, uint64_t window_size);
+    extern void calculate_wma(double * input, double * output, uint64_t data_size, uint64_t window_size);
 #if defined(__cplusplus) && (! defined(__ISPC_NO_EXTERN_C) || !__ISPC_NO_EXTERN_C )
 } /* end extern C */
 #endif // __cplusplus
